feat(x/scheduler): order task types through dependency graph
